% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iobed.video-package.R
\docType{package}
\name{iobed.video-package}
\alias{iobed.video}
\alias{iobed.video-package}
\title{iobed.video: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UBESP-DCTV/iobed.video}
  \item Report bugs at \url{https://github.com/UBESP-DCTV/iobed.video/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@gmail.com} (\href{https://orcid.org/0000-0002-0520-7428}{ORCID}) [copyright holder]

}
\keyword{internal}
